Refactor views/board&component
