+++ lib/screens/Board.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:helloflutter/screens/board/BoardQnaList.dart';
-import 'package:helloflutter/screens/board/BoardJobFindingList.dart';
+import 'package:helloflutter/views/Board/BoardQnaList.dart';
+import 'package:helloflutter/views/Board/BoardJobFindingList.dart';
 import 'package:helloflutter/models/YrkData.dart';
-import 'package:helloflutter/views/YrkCardView.dart';
-import 'package:helloflutter/views/YrkWidgetHeaderView.dart';
-import 'package:helloflutter/views/YrkPageView.dart';
+import 'package:helloflutter/views/components/YrkCardView.dart';
+import 'package:helloflutter/views/components/YrkWidgetHeaderView.dart';
+import 'package:helloflutter/views/components/YrkPageView.dart';
 
 class Board extends StatefulWidget {
   Board({@required this.onPushNavigator});
