+++ lib/screens/Signin.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import 'package:helloflutter/views/components/YrkTextField.dart';
-import 'package:helloflutter/views/components/YrkTextStyle.dart';
-import 'package:helloflutter/views/components/YrkButton.dart';
+import 'package:helloflutter/views/widgets/YrkTextField.dart';
+import 'package:helloflutter/views/widgets/YrkTextStyle.dart';
+import 'package:helloflutter/views/widgets/YrkButton.dart';
 
 class SignIn extends StatefulWidget {
   @override
