+++ lib/views/board/BoardJobFindingList.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:helloflutter/navigator/DetailPage.dart';
 import 'package:helloflutter/navigator/TabItem.dart';
 import 'package:helloflutter/models/YrkData.dart';
-import 'package:helloflutter/views/YrkListView.dart';
+import 'package:helloflutter/views/components/YrkListView.dart';
 
 class BoardJobFindingList extends YrkListView {
   BoardJobFindingList({
