+++ lib/views/components/YrkCardView.dart
