+++ lib/views/components/YrkTabBarView.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:helloflutter/views/components/YrkDotsIndicator.dart';
+import 'package:helloflutter/views/widgets/YrkDotsIndicator.dart';
 
 class YrkTabBarView extends StatelessWidget {
   YrkTabBarView({
